simplify resolution
