+++ crates/datasec/src/did/web.rs
@@ -7,7 +7,6 @@
 //! - <https://w3c-ccg.github.io/did-method-web>
 //! - <https://w3c.github.io/did-resolution>
 
-pub mod operator;
 pub mod resolver;
 
 #[allow(clippy::module_name_repetitions)]
