+++ crates/datasec/src/jose/jwk.rs
@@ -5,6 +5,21 @@
 //!
 //! See [RFC7517] for more detail.
 //!
+//! TODO:
+//! Support:
+//! (key) type: EcdsaSecp256k1VerificationKey2019 | JsonWebKey2020 | Ed25519VerificationKey2020 | Ed25519VerificationKey2018 | X25519KeyAgreementKey2019
+//! crv: Ed25519 | secp256k1 | P-256 | P-384 | p-521
+//!
+//! JWK Thumbprint [RFC7638]
+//! It is RECOMMENDED that JWK kid values are set to the public key fingerprint:
+//!  - create SHA-256 hash of UTF-8 representation of JSON from {crv,kty,x,y}
+//!
+//! For example:
+//!  - JSON: {"crv":"Ed25519","kty":"OKP","x":"11qYAYKxCrfVS_7TyWQHOg7hcvPapiMlrwIaaPcHURo"}
+//!  - SHA-256: 90facafea9b1556698540f70c0117a22ea37bd5cf3ed3c47093c1707282b4b89
+//!  - base64url JWK Thumbprint: "kPrK_qmxVWaYVA9wwBF6Iuo3vVzz7TxHCTwXBygrS4k"
+//!
+//! [RFC7638]: https://www.rfc-editor.org/rfc/rfc7638
 //! [RFC7517]: https://www.rfc-editor.org/rfc/rfc7517
 
 use serde::{Deserialize, Serialize};
